Nuplayer: Update request input buffer delay as per fps
Currently, we have static 10 ms delay to fetch back the
decoder input buffers for all usecases when the queue is
empty. For HFR usecases (>480 fps), fetching back the
input buffers after 10ms is leading to the stalls in
input buffer pipeline and so the performance issue is
observed. Hence update the fetch delay as a function
of operating rate to speed up the pipeline delay.

Bug: b/208475704
Change-Id: I95163d3d05721cc0a12ee9f486a7d3a3b8b8a600
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: ReveRTX <rev9rt@gmail.com>